examples/parse_pcap/requirements.txt
scapy==2.5.0
sqlalchemy>=2.0.0
python-dotenv==1.0.0
click==8.1.7
tqdm>=4.66.3
